shared state implemented.
Additionally fixed the problem with guards. For reads serialization must happen immediately upon read otherwise there has to be a copy of the resp type out of shared state. This breaks modularity but is necessary to get rid of an unecessary copy.

All entries into shared state involve a copy. All reads involve a copy at the serialization step.